requirements.txt
Django>=5.2,<5.3
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-filter>=24.0
django-cors-headers>=4.3
psycopg[binary]>=3.1
Pillow>=10.0
gunicorn>=22.0
